Added online paradigm
The Feedback is a bar with 4 rectangulars, which represents the probabilities of 4 classes. During the MI phase, the bar is shown and is updated with every frame and the length of the class rectangular represents the probability of this class (ie small length represents low probability and longerr length higher probability.